refactor: ビジネスロジックを domain/ サブドメインに抽出（フェーズ 2）
- domain/game/: event-generator.ts, sprint.ts（createEvents, createSprintSummary）
- domain/quiz/: question-picker.ts, answer-evaluator.ts, combo-calculator.ts, tag-stats.ts, study-question-pool.ts
- domain/scoring/: difficulty.ts
- domain/team/: team-classifier.ts
- domain/achievement/: achievement-checker.ts
- ランダム依存関数に randomFn 引数を追加して純粋化（pickQuestion, createEvents, shuffleArray）
- scoring-types.ts を team-types.ts, achievement-types.ts, scoring-types.ts に分配
- 旧ファイルは後方互換用の再エクスポートとして維持
- 全33テストスイート/598テストパス確認
