+++ src/features/agile-quiz-sugoroku/domain/achievement/index.ts
@@ -1,0 +1,4 @@
+/**
+ * achievement サブドメイン - 再エクスポート
+ */
+export { ACHIEVEMENTS, checkAchievements } from './achievement-checker';
