+++ src/features/agile-quiz-sugoroku/domain/game/index.ts
@@ -1,0 +1,5 @@
+/**
+ * game サブドメイン - 再エクスポート
+ */
+export { createEvents, makeEvents } from './event-generator';
+export { createSprintSummary } from './sprint';
