+++ src/features/agile-quiz-sugoroku/domain/scoring/index.ts
@@ -1,0 +1,8 @@
+/**
+ * scoring サブドメイン - 再エクスポート
+ */
+export {
+  DIFFICULTY_CONFIGS,
+  getDifficultyConfig,
+  calculateGradeWithDifficulty,
+} from './difficulty';
