+++ src/features/agile-quiz-sugoroku/domain/team/index.ts
@@ -1,0 +1,4 @@
+/**
+ * team サブドメイン - 再エクスポート
+ */
+export { classifyEngineerType } from './team-classifier';
