+++ src/features/agile-quiz-sugoroku/domain/team/team-classifier.ts
@@ -1,0 +1,12 @@
+/**
+ * エンジニアタイプ分類
+ *
+ * 旧 engineer-classifier.ts から移動。
+ */
+import { EngineerType, ClassifyStats } from '../types';
+import { ENGINEER_TYPES } from '../../constants';
+
+/** エンジニアタイプを判定 */
+export function classifyEngineerType(data: ClassifyStats): EngineerType {
+  return ENGINEER_TYPES.find((t) => t.condition(data)) ?? ENGINEER_TYPES[ENGINEER_TYPES.length - 1];
+}
